moved launching utils to internal utils package
