+++ launching/src/main/java/org/jboss/tools/ssp/internal/launching/util/LibraryLocationUtils.java
@@ -6,7 +6,7 @@
  * 
  * Contributors: Red Hat, Inc.
  ******************************************************************************/
-package org.jboss.tools.ssp.launching;
+package org.jboss.tools.ssp.internal.launching.util;
 
 import java.io.File;
 import java.net.URL;
