+++ server-wildfly/src/main/java/org/jboss/tools/ssp/server/wildfly/servertype/impl/JBossStopLauncher.java
@@ -20,7 +20,7 @@ import org.jboss.tools.ssp.eclipse.debug.core.Launch;
 import org.jboss.tools.ssp.eclipse.jdt.launching.ExecutionArguments;
 import org.jboss.tools.ssp.eclipse.jdt.launching.IVMRunner;
 import org.jboss.tools.ssp.eclipse.jdt.launching.VMRunnerConfiguration;
-import org.jboss.tools.ssp.launching.NativeEnvironmentUtil;
+import org.jboss.tools.ssp.internal.launching.util.NativeEnvironmentUtils;
 
 public class JBossStopLauncher {
 	private JBossServerDelegate delegate;
@@ -133,7 +133,7 @@ public class JBossStopLauncher {
 	
 	private String[] getEnvironment(boolean appendNativeEnv) {
 		Map<String, String> configEnv = getEnvironmentFromServer();
-		return NativeEnvironmentUtil.getDefault().getEnvironment(configEnv, appendNativeEnv);
+		return NativeEnvironmentUtils.getDefault().getEnvironment(configEnv, appendNativeEnv);
 	}
 
 	private Map<String, String>  getEnvironmentFromServer() {
